src/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2
django-countries==7.3.2
django-debug-toolbar==3.4.0
django-filter==22.1
django-redis==5.2.0
django-summernote==0.8.20.0
django-taggit==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Faker==14.2.0
importlib-metadata==4.12.0
kombu==5.2.4
Markdown==3.4.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.1
